+++ previous-versions/assess/src/sdk/excelreports.ts
@@ -116,7 +116,7 @@ export class ExcelReports {
                     );
                 }
                 break;
-            case [401, 404].includes(httpRes?.status):
+            case [400, 401, 402, 403, 404, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),
@@ -227,7 +227,7 @@ export class ExcelReports {
                     );
                 }
                 break;
-            case [401, 404].includes(httpRes?.status):
+            case [400, 401, 402, 403, 404, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),
@@ -340,7 +340,7 @@ export class ExcelReports {
                     );
                 }
                 break;
-            case [401, 404].includes(httpRes?.status):
+            case [400, 401, 402, 403, 404, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),
@@ -452,7 +452,7 @@ export class ExcelReports {
                     );
                 }
                 break;
-            case [401, 404].includes(httpRes?.status):
+            case [400, 401, 402, 403, 404, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),