+++ previous-versions/assess/src/sdk/models/operations/getaccountsforenhancedbalancesheet.ts
@@ -40,7 +40,7 @@ export class GetAccountsForEnhancedBalanceSheetResponse extends SpeakeasyBase {
     enhancedReport?: shared.EnhancedReport;
 
     /**
-     * Your API request was not properly authorized.
+     * The request made is not valid.
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;