+++ previous-versions/assess/src/sdk/models/operations/getcommercerevenuemetrics.ts
@@ -58,7 +58,7 @@ export class GetCommerceRevenueMetricsResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
-     * Your API request was not properly authorized.
+     * The request made is not valid.
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;