+++ previous-versions/assess/src/sdk/models/operations/getdataintegritysummaries.ts
@@ -34,7 +34,7 @@ export class GetDataIntegritySummariesResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
-     * Your API request was not properly authorized.
+     * Your `query` parameter was not correctly formed
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;