+++ previous-versions/assess/src/sdk/models/operations/getexcelreportgenerationstatus.ts
@@ -28,7 +28,7 @@ export class GetExcelReportGenerationStatusResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
-     * Your API request was not properly authorized.
+     * Your `query` parameter was not correctly formed
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;