+++ previous-versions/assess/RELEASES.md
@@ -627,3 +627,13 @@ Based on:
 - [typescript v0.36.0] previous-versions/assess
 ### Releases
 - [NPM v0.36.0] https://www.npmjs.com/package/@codat/assess/v/0.36.0 - previous-versions/assess
+
+## 2023-10-30 01:23:36
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.37.0] previous-versions/assess
+### Releases
+- [NPM v0.37.0] https://www.npmjs.com/package/@codat/assess/v/0.37.0 - previous-versions/assess